requirements.txt
streamlit
google-generativeai
google-genai
python-dotenv
Pillow
fal-client
